Create util for currentNetwork Display currentNetwork in warning banner
